Split sandbox profile builders
